save files now existgit add -A

+++ src/util/math.py
@@ -2,5 +2,5 @@ from pygame import math
 
 Vector2 = math.Vector2
 
-def clamp(min: int, value: int, max: int) -> int:
+def clamp(min: float, value: float, max: float) -> float:
     return sorted([min, value, max])[1] # always returns middle value
